Move Foundation.Infrastructure to new PersistenceInfrastructure project
