+++ account-management/src/Foundation/Foundation.csproj
@@ -23,11 +23,11 @@
             <PrivateAssets>all</PrivateAssets>
             <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
         </PackageReference>
-        <PackageReference Include="Scrutor" Version="4.2.2"/>
     </ItemGroup>
 
     <ItemGroup>
         <ProjectReference Include="..\..\..\backend-foundation\src\DddCqrsFramework\DddCqrsFramework.csproj"/>
+        <ProjectReference Include="..\..\..\backend-foundation\src\PersistenceInfrastructure\PersistenceInfrastructure.csproj"/>
     </ItemGroup>
 
 </Project>
