+++ account-management/src/WebApi/Program.cs
@@ -2,7 +2,7 @@ using PlatformPlatform.AccountManagement.Application;
 using PlatformPlatform.AccountManagement.Infrastructure;
 using PlatformPlatform.AccountManagement.WebApi;
 using PlatformPlatform.AccountManagement.WebApi.Endpoints;
-using PlatformPlatform.Foundation;
+using PlatformPlatform.Foundation.PersistenceInfrastructure;
 using PlatformPlatform.Foundation.WebApi;
 
 var builder = WebApplication.CreateBuilder(args);
